mark each relevant file or folder used by AFCache as exclude from backup to prevent automatic deletion by system
for exclusion from iCloud backup it is sufficient to mark a folder for exclusion, but for cache purging all files must be marked.
this is not documented but validated by tests on an iPhone5C where videos were uploaded until it is almost full (~1GB free space) then loading
some issues until it is full (~10mb free) then taking videos with that camera to trigger cache purging.
-
https://redmine.purplemotion.com/issues/5661 (refs #5661)